Remove unused assert code from a test
